add 1 files
In this project , We Create a JavaScript Form Validation Project . the code that includes HTML, CSS (Bootstrap), and JavaScript to implement the form validation based on the provided criteria. This code snippet includes the HTML form structure with Bootstrap styling classes. The JavaScript code adds an event listener to the form's submit event and performs the necessary validations for each field based on the given criteria. If any validation criteria are not met, an error message is displayed next to the corresponding field. If there are no error messages, the form is considered valid and can be submitted. The clearErrors() function clears any previous error messages before performing the validations. The checkErrors() function checks if there are any error messages displayed, and if so, it returns true, indicating that the form should not be submitted.